requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
jinja2==3.1.4
SQLAlchemy==2.0.34
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3